Refactoring Code IV

+++ src/models/User.js
@@ -16,25 +16,21 @@ const User = {
   },
 
   //  * CREER UN UTILISATEUR
-  createUser: ({ name, email, role, password }) => {
-    const hashedPassword = bcrypt.hashSync(password, 9);
+  createUser: ({ name, email, role, password }, finalQuota) => {
+    const hashedPassword = bcrypt.hashSync(password, 10);
     const stmt = db.prepare(`
-      INSERT INTO users (id,name, email, role, password)
-      VALUES (?, ?, ?, ?,?)
+      INSERT INTO users (name, email, role, password, quota)
+      VALUES (?, ?, ?,?, ?)
     `);
 
-    const result = stmt.run(
-      uuidv4(),
-      name,
-      email,
-      role || "user",
-      hashedPassword,
-    );
+    const result = stmt.run(name, email, "user", hashedPassword, finalQuota);
     return {
+      id: result.lastInsertRowid,
       name,
       email,
       role,
       password,
+      quota: finalQuota,
     };
   },
 
@@ -45,14 +41,51 @@ const User = {
 
   //  * UPDATE USER
 
-  update: (id, { name, email, password }) => {
-    const stmt = db.prepare(`UPDATE users SET name=?, email=?, WHERE id=?`);
-    return stmt.run(id, name, email);
+  updateRow: (id, { name, email, password, quota, used_space }) => {
+    const fields = [];
+    const values = [];
+
+    if (name) {
+      fields.push("name = ?");
+      values.push(name);
+    }
+    if (email) {
+      fields.push("email = ?");
+      values.push(email);
+    }
+    if (password) {
+      fields.push("password = ?");
+      values.push(password);
+    }
+    if (quota !== undefined) {
+      fields.push("quota = ?");
+      values.push(quota);
+    }
+
+    if (used_space) {
+      fields.push("used_space=?");
+      values.push(used_space);
+    }
+
+    if (fields.length === 0) {
+      return null;
+    }
+
+    const stmt = db.prepare(
+      `UPDATE users SET ${fields.join(", ")} WHERE id=? AND role != 'admin'`,
+    );
+    values.push(id);
+    return stmt.run(...values);
   },
 
   delete: (id) => {
-    const stmt = db.prepare(`DELETE FROM users WHERE id=?`);
+    const stmt = db.prepare(`DELETE FROM users WHERE id=? AND role !='admin'`);
     return stmt.run(id);
+  },
+
+  getAllUsers: () => {
+    const stmt = db.prepare("SELECT * FROM users ORDER BY created_at DESC");
+    return stmt.all();
   },
 };
